Restructure commands to make release creation independent
